+++ src/operators/reduction.cpp
@@ -10,7 +10,7 @@ namespace md{
             return op::apply_reduction<op::Sum>(this, node, axes);
         }
 
-        Node GraphInternal::sum(Node node, short axis){
+        Node GraphInternal::sum(Node node, int axis){
             auto a = auto_infer_axes(node);
             if(axis == auto_infer){
                 return op::apply_reduction<op::Sum>(this, node, auto_infer_axes(node));
@@ -22,7 +22,7 @@ namespace md{
             return op::apply_reduction<op::Mean>(this, node, axes);
         }
 
-        Node GraphInternal::mean(Node node, short axis){
+        Node GraphInternal::mean(Node node, int axis){
             if(axis == auto_infer){
                 return op::apply_reduction<op::Mean>(this, node, auto_infer_axes(node));
             }
@@ -33,7 +33,7 @@ namespace md{
             return op::apply_reduction<op::Product>(this, node, axes);
         }
 
-        Node GraphInternal::prod(Node node, short axis){
+        Node GraphInternal::prod(Node node, int axis){
             if(axis == auto_infer){
                 return op::apply_reduction<op::Product>(this, node, auto_infer_axes(node));
             }
@@ -44,7 +44,7 @@ namespace md{
             return op::apply_reduction<op::AllTrue>(this, node, axes);
         }
 
-        Node GraphInternal::all_true(Node node, short axis){
+        Node GraphInternal::all_true(Node node, int axis){
             if(axis == auto_infer){
                 return op::apply_reduction<op::AllTrue>(this, node, auto_infer_axes(node));
             }
@@ -55,7 +55,7 @@ namespace md{
             return op::apply_reduction<op::AnyTrue>(this, node, axes);
         }
 
-        Node GraphInternal::any_true(Node node, short axis){
+        Node GraphInternal::any_true(Node node, int axis){
             if(axis == auto_infer){
                 return op::apply_reduction<op::AnyTrue>(this, node, auto_infer_axes(node));
             }
